modified:   .DS_Store 	new file:   .clangd 	modified:   .vscode/settings.json 	modified:   main/.DS_Store 	modified:   main/CMakeLists.txt 	modified:   main/lv_conf.h 	modified:   main/main.c 	modified:   main/odometer/odometer.c 	modified:   main/odometer/odometer.h 	deleted:    main/tach_ui/.DS_Store 	deleted:    main/tach_ui/CMakeLists.txt 	deleted:    main/tach_ui/components/ui_comp_hook.c 	deleted:    main/tach_ui/filelist.txt 	deleted:    main/tach_ui/fonts/school_bell_48.c 	deleted:    main/tach_ui/fonts/ui_font_DotoLarge.c 	deleted:    main/tach_ui/fonts/ui_font_Doto_48.c 	deleted:    main/tach_ui/images/ui_img_1656279599.c 	deleted:    main/tach_ui/images/ui_img_sti_logo_png.c 	deleted:    main/tach_ui/project.info 	deleted:    main/tach_ui/screens/ui_Screen1.c 	deleted:    main/tach_ui/screens/ui_Screen1.h 	deleted:    main/tach_ui/ui.c 	deleted:    main/tach_ui/ui.h 	deleted:    main/tach_ui/ui_events.h 	deleted:    main/tach_ui/ui_helpers.c 	deleted:    main/tach_ui/ui_helpers.h 	renamed:    main/tach_ui/boot/boot_screen.c -> main/ui/boot/boot_screen.c 	renamed:    main/tach_ui/boot/boot_screen.h -> main/ui/boot/boot_screen.h 	modified:   main/ui/images/ui_img_gauge_background_png.c 	modified:   main/ui/project.info 	modified:   main/ui/screens/ui_MainSpeedo.c 	modified:   main/ui/screens/ui_MainSpeedo.h 	modified:   main/ui/ui.h 	modified:   sdkconfig

+++ main/odometer/odometer.h
@@ -19,4 +19,4 @@ void odometer_force_save(void);
 uint64_t odometer_get_meters(void);
 
 // Total distance converted to miles for display.
-double odometer_get_miles(void);
+int odometer_get_miles(void);
